project-app/backend/requirements.txt
flask
flask-cors
PyPDF2
ollama
werkzeug
python-dotenv